change version to an int so it can index into an array.
I the future if the versions become sparse or large we will create a "proxy"
version index by putting the standardized indexes into an enum and using
the enum ordinal instead.

+++ ethereum/core/src/main/java/tech/pegasys/pantheon/ethereum/chain/GenesisState.java
@@ -221,7 +221,7 @@ public final class GenesisState {
     final Wei balance;
     final Map<UInt256, UInt256> storage;
     final BytesValue code;
-    final long version;
+    final int version;
 
     static GenesisAccount fromAllocation(final GenesisAllocation allocation) {
       return new GenesisAccount(
@@ -241,7 +241,7 @@ public final class GenesisState {
       this.address = withNiceErrorMessage("address", hexAddress, Address::fromHexString);
       this.balance = withNiceErrorMessage("balance", balance, this::parseBalance);
       this.code = hexCode != null ? BytesValue.fromHexString(hexCode) : null;
-      this.version = version != null ? Long.decode(version) : Account.DEFAULT_VERSION;
+      this.version = version != null ? Integer.decode(version) : Account.DEFAULT_VERSION;
       this.storage = parseStorage(storage);
     }
